+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 VMware Aria Operations Integration SDK
 --------------------------------------
 ## Unreleased
+* Default container registry can now be specified in global config file
 * Adds two flags to mp-test for controlling collection parameters:
   * '--collection-number': Sets the collection number. This is useful for testing functionality that only occurs on some collection cycles.
   * '--collection-window-duration': Sets the duration of the collection window. This is useful for testing how the adapter behaves for longer or shorter collection windows.